+++ ThirdPartyAdapters/inmobi/inmobi/src/main/java/com/google/ads/mediation/inmobi/InMobiInterstitialWrapper.java
@@ -2,6 +2,7 @@ package com.google.ads.mediation.inmobi;
 
 import androidx.annotation.VisibleForTesting;
 import com.inmobi.ads.InMobiInterstitial;
+import com.inmobi.ads.WatermarkData;
 import java.util.Map;
 
 /** Wrapper around InMobiInterstitial class. */
@@ -24,6 +25,10 @@ public class InMobiInterstitialWrapper {
 
   public void setKeywords(final String keywords) {
     inMobiInterstitial.setKeywords(keywords);
+  }
+
+  public void setWatermarkData(final WatermarkData watermarkData) {
+    inMobiInterstitial.setWatermarkData(watermarkData);
   }
 
   public void load() {