+++ ThirdPartyAdapters/inmobi/inmobi/src/main/java/com/google/ads/mediation/inmobi/renderers/InMobiBannerAd.java
@@ -15,10 +15,12 @@
 package com.google.ads.mediation.inmobi.renderers;
 
 import static com.google.ads.mediation.inmobi.InMobiConstants.ERROR_BANNER_SIZE_MISMATCH;
+import static com.google.ads.mediation.inmobi.InMobiConstants.WATERMARK_ALPHA;
 import static com.google.ads.mediation.inmobi.InMobiMediationAdapter.TAG;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 import android.widget.FrameLayout;
@@ -39,6 +41,7 @@ import com.google.android.gms.ads.mediation.MediationBannerAdConfiguration;
 import com.inmobi.ads.AdMetaInfo;
 import com.inmobi.ads.InMobiAdRequestStatus;
 import com.inmobi.ads.InMobiBanner;
+import com.inmobi.ads.WatermarkData;
 import com.inmobi.ads.listeners.BannerAdEventListener;
 import java.util.Map;
 
@@ -127,6 +130,11 @@ public abstract class InMobiBannerAd extends BannerAdEventListener implements Me
     inMobiBannerWrapper.setAnimationType(InMobiBanner.AnimationType.ANIMATION_OFF);
 
     inMobiBannerWrapper.setListener(InMobiBannerAd.this);
+
+    String watermark = mediationBannerAdConfiguration.getWatermark();
+    if (!TextUtils.isEmpty(watermark)) {
+      inMobiBannerWrapper.setWatermarkData(new WatermarkData(watermark, WATERMARK_ALPHA));
+    }
 
     /*
      * Wrap InMobi's ad view to limit the dependency on its methods. For example, the method