+++ ThirdPartyAdapters/inmobi/inmobi/src/main/java/com/google/ads/mediation/inmobi/renderers/InMobiInterstitialAd.java
@@ -16,10 +16,12 @@ package com.google.ads.mediation.inmobi.renderers;
 
 import static com.google.ads.mediation.inmobi.InMobiConstants.ERROR_AD_DISPLAY_FAILED;
 import static com.google.ads.mediation.inmobi.InMobiConstants.ERROR_AD_NOT_READY;
+import static com.google.ads.mediation.inmobi.InMobiConstants.WATERMARK_ALPHA;
 import static com.google.ads.mediation.inmobi.InMobiMediationAdapter.TAG;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.text.TextUtils;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
@@ -36,6 +38,7 @@ import com.google.android.gms.ads.mediation.MediationInterstitialAdConfiguration
 import com.inmobi.ads.AdMetaInfo;
 import com.inmobi.ads.InMobiAdRequestStatus;
 import com.inmobi.ads.InMobiInterstitial;
+import com.inmobi.ads.WatermarkData;
 import com.inmobi.ads.listeners.InterstitialAdEventListener;
 import java.util.Map;
 
@@ -104,6 +107,12 @@ public abstract class InMobiInterstitialAd extends InterstitialAdEventListener
 
     InMobiAdapterUtils.configureGlobalTargeting(
         mediationInterstitialAdConfiguration.getMediationExtras());
+
+    String watermark = mediationInterstitialAdConfiguration.getWatermark();
+    if (!TextUtils.isEmpty(watermark)) {
+      inMobiInterstitialWrapper.setWatermarkData(new WatermarkData(watermark, WATERMARK_ALPHA));
+    }
+
     internalLoadAd(inMobiInterstitialWrapper);
   }