+++ ThirdPartyAdapters/inmobi/inmobi/src/main/java/com/google/ads/mediation/inmobi/renderers/InMobiRewardedAd.java
@@ -16,6 +16,7 @@ package com.google.ads.mediation.inmobi.renderers;
 
 import static com.google.ads.mediation.inmobi.InMobiConstants.ERROR_AD_DISPLAY_FAILED;
 import static com.google.ads.mediation.inmobi.InMobiConstants.ERROR_AD_NOT_READY;
+import static com.google.ads.mediation.inmobi.InMobiConstants.WATERMARK_ALPHA;
 import static com.google.ads.mediation.inmobi.InMobiMediationAdapter.TAG;
 
 import android.content.Context;
@@ -39,6 +40,7 @@ import com.google.android.gms.ads.rewarded.RewardItem;
 import com.inmobi.ads.AdMetaInfo;
 import com.inmobi.ads.InMobiAdRequestStatus;
 import com.inmobi.ads.InMobiInterstitial;
+import com.inmobi.ads.WatermarkData;
 import com.inmobi.ads.listeners.InterstitialAdEventListener;
 import java.util.Map;
 
@@ -127,6 +129,12 @@ public abstract class InMobiRewardedAd extends InterstitialAdEventListener
 
     InMobiAdapterUtils.configureGlobalTargeting(
         mediationRewardedAdConfiguration.getMediationExtras());
+
+    String watermark = mediationRewardedAdConfiguration.getWatermark();
+    if (!TextUtils.isEmpty(watermark)) {
+      inMobiRewardedAdWrapper.setWatermarkData(new WatermarkData(watermark, WATERMARK_ALPHA));
+    }
+
     internalLoadAd(inMobiRewardedAdWrapper);
   }
   // endregion